+++ nacc_attribute_deriver/attributes/nacc/imaging/scan.py
@@ -19,13 +19,8 @@ class NACCSCANAttributeCollection(AttributeCollection):
 
     def _create_scan_mri_dates(self) -> str:
         """Gets the date of the MRI scan - temporary derived variable."""
-        # TODO: either studydate (mridashboard) or scandt (ucdmrisbm)
-        # need to confirm how the data is represented in the table
-        # do mridashboard then ucdmrisbm
         self.__scan.assert_required(REQUIRED_FIELDS[MRIPrefix.SCAN_MRI_QC])
         scandate = self.__scan.get_value("study_date")
-        # if not scandate:
-        #     scandate = self.get_mri_value('scandt', MRIPrefix.MRI_SBM)
 
         scandate = datetime_from_form_date(scandate)
 
@@ -36,14 +31,8 @@ class NACCSCANAttributeCollection(AttributeCollection):
 
     def _create_scan_pet_dates(self) -> str:
         """Gets the date of the PET scan - temporary derived variable"""
-        # TODO: either or scan_date(petdashboard) scandate (berkeley files)
-        # need to confirm how the data is represented in the table
-        # try petdashboard first, then berkeley files
         self.__scan.assert_required(REQUIRED_FIELDS[PETPrefix.SCAN_PET_QC])
         scandate = self.__scan.get_value("scan_date")
-        # if not scandate:
-        #     for subprefix in PETPrefix.analysis_files():
-        #         scandate = self.get_pet_value('scandate', subprefix)
 
         scandate = datetime_from_form_date(scandate)
         if not scandate: