+++ nacc_attribute_deriver/attributes/nacc/modules/__init__.py
@@ -1,4 +1,6 @@
 # ruff: noqa
 from .cross_module import *
 from .np import *
+from .mds import *
+from .milestone import *
 from .uds import *