+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 Documentation of release versions of `nacc-attribute-deriver`
 
+## 1.2.2
+
+* Adds ability for UDS NACC derived variables to account for follow-up forms
+    * If a raw value is not specified, checks the subject-level derived value as well before setting to the missing/null case
+    * Adds rules to the config for UDS NACC derived variables to write to `subject.info.derived`
+    * Updates the following 9 variables to support this:
+        * Form A1: `naccnihr`
+        * Form D1: `naccalzp`, `nacclbde`, `nacclbdp`, `naccudsd`, `naccetpr`, `naccppa`, `naccbvft`, `nacclbds`
+* Updates `uds_primary_language` rule to return `None` (handles case where subsequent visits do not re-specify the value)
+* Fixes `total_uds_visits` rule
+* Fixes missing import for `mds` and `milestone` NACC rules
+
+## 1.2.1
+
+* Allows for `subject.info.derived.np_death_age` to be unset when verifying existence of NP data
+
+## 1.2.0
+
+* Includes MQT V2.1 attributes over NP data
+
 ## 1.1.0
 
 * Refines MQT V2 variables, namely making sure code aligns with the FW data model